clean up import preview row cap
